src/server/pyproject.toml
[project]
name = "Opogen"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Equipo45", email = "ismiruiz@gmail.com"},
]
dependencies = [
    "PyPDF2>=3.0.1",
    "python-docx>=1.1.0",
    "langchain>=0.1.14",
    "langchain-openai>=0.1.1",
    "markdown>=3.6",
    "fpdf>=1.7.2",
    "Flask==2.2.3",
    "Flask-Cors==3.0.10",
    "Werkzeug==2.2.2",
    "reportlab>=4.1.0",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    # "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    # "I",
]
ignore = [
    # star imports
    "F405", "F403",
    # lambda expression assignment
    "E731",
    # Loop control variable not used within loop body
    "B007",
    "E501",  # Ignore "Line too long" error
]

[tool.codespell]
count = true
quiet-level = 3

[tool.isort]
use_parentheses = true
skip_gitignore = true